refactor(profile_plugin): split frontend APIs by UI concern
Move HTTP handlers out of the ProfilePlugin god-class into api/* mixins:
static assets/config, runs/tools, hosts, tool data, capture, and cache
warm-up. plugin.py keeps construction + route registration only (~280 LOC).
All 113 package tests pass.
