+++ plugin/xprof/profile_plugin/api/types.py
@@ -1,0 +1,8 @@
+# Copyright 2026 The XProf Authors. All Rights Reserved.
+"""Shared TypedDicts for frontend API responses."""
+
+from __future__ import annotations
+
+from typing import TypedDict
+
+HostMetadata = TypedDict('HostMetadata', {'hostname': str})
